[FLINK-19770][python][tests] Extract PythonProgramOptionTest#testConfigurePythonExecution into an ITCase
This closes #13756.
