+++ flink-python/src/test/java/org/apache/flink/client/cli/PythonProgramOptionsTest.java
@@ -18,8 +18,6 @@
 
 package org.apache.flink.client.cli;
 
-import org.apache.flink.client.program.PackagedProgram;
-import org.apache.flink.client.program.ProgramInvocationException;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.python.PythonOptions;
 
@@ -27,15 +25,6 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Options;
 import org.junit.Before;
 import org.junit.Test;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.FileSystems;
-import java.nio.file.FileVisitResult;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 
 import static org.apache.flink.client.cli.CliFrontendParser.PYARCHIVE_OPTION;
 import static org.apache.flink.client.cli.CliFrontendParser.PYEXEC_OPTION;
@@ -49,10 +38,9 @@ import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 
 /**
- * Tests for {@link PythonProgramOptions}.
+ * Test for {@link PythonProgramOptions}.
  */
 public class PythonProgramOptionsTest {
-
 	private Options options;
 
 	@Before
@@ -114,45 +102,4 @@ public class PythonProgramOptionsTest {
 			new String[] {"--python", "xxx.py", "--pyModule", "xxx", "userarg1", "userarg2"},
 			programOptions.getProgramArgs());
 	}
-
-	@Test
-	public void testConfigurePythonExecution() throws IllegalAccessException, NoSuchFieldException, CliArgsException, ProgramInvocationException, IOException {
-		final String[] args = {
-			"--python", "xxx.py",
-			"--pyModule", "xxx",
-			"--pyFiles", "/absolute/a.py,relative/b.py,relative/c.py",
-			"--pyRequirements", "d.txt#e_dir",
-			"--pyExecutable", "/usr/bin/python",
-			"--pyArchives", "g.zip,h.zip#data,h.zip#data2",
-			"userarg1", "userarg2"
-		};
-
-		final File[] dummyJobJar = {null};
-		Files.walkFileTree(FileSystems.getDefault().getPath("target/dummy-job-jar"), new SimpleFileVisitor<Path>() {
-			@Override
-			public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-				FileVisitResult result = super.visitFile(file, attrs);
-				if (file.getFileName().toString().startsWith("flink-python")) {
-					dummyJobJar[0] = file.toFile();
-				}
-				return result;
-			}
-		});
-
-		PackagedProgram packagedProgram = PackagedProgram.newBuilder()
-			.setArguments(args).setJarFile(dummyJobJar[0])
-			.build();
-
-		Configuration configuration = new Configuration();
-		ProgramOptionsUtils.configurePythonExecution(configuration, packagedProgram);
-
-		assertEquals("/absolute/a.py,relative/b.py,relative/c.py", configuration.get(PythonOptions.PYTHON_FILES));
-		assertEquals("d.txt#e_dir", configuration.get(PYTHON_REQUIREMENTS));
-		assertEquals("g.zip,h.zip#data,h.zip#data2", configuration.get(PythonOptions.PYTHON_ARCHIVES));
-		assertEquals("/usr/bin/python", configuration.get(PYTHON_EXECUTABLE));
-		assertArrayEquals(
-			new String[] {"--python", "xxx.py", "--pyModule", "xxx", "userarg1", "userarg2"},
-			packagedProgram.getArguments());
-	}
-
 }
